requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
python-dotenv==0.21.1
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13